2025AIE_prj09_slack_bot_monitoring/requirements.txt
slack_sdk
boto3
pyyaml